# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the VyoLayer API documentation.",
    "title": "VyoLayer API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "API Support",
      "url": "http://www.swagger.io/support",
      "email": "support@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "host": "localhost:6999",
  "basePath": "/api/v1",
  "paths": {
...
    "/organizations": {
      "get": {
        "description": "List organizations the user is a member of.",
        "produces": ["application/json"],
        "tags": ["Organizations"],
        "summary": "List organizations",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.SuccessResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/dto.OrganizationDTO"
                      }
                    }
                  }
                }
              ]
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      },
      "post": {
        "description": "Create a new organization. The authenticated user becomes the owner.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Organizations"],
        "summary": "Create organization",
        "parameters": [
          {
            "description": "Organization details",
            "name": "organization",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.CreateOrganizationRequestDTO"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.SuccessResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/dto.OrganizationResponseDTO"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      }
    },
...
    "/organizations/onboarding": {
      "post": {
        "description": "Guided onboarding flow for creating a new organization.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Organizations"],
        "summary": "Onboard organization",
        "parameters": [
          {
            "description": "Organization details",
            "name": "organization",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.CreateOrganizationRequestDTO"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.SuccessResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/dto.OrganizationResponseDTO"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      }
    },
    "/organizations/slug/{slug}": {
      "get": {
        "description": "Retrieve organization details by slug. User must be a member.",
        "produces": ["application/json"],
        "tags": ["Organizations"],
        "summary": "Get organization by slug",
        "parameters": [
          {
            "type": "string",
            "description": "Organization slug",
            "name": "slug",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.SuccessResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/dto.OrganizationResponseDTO"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      }
    },
    "/organizations/{id}": {
      "get": {
        "description": "Retrieve organization details by ID. User must be a member.",
        "produces": ["application/json"],
        "tags": ["Organizations"],
        "summary": "Get organization by ID",
        "parameters": [
          {
            "type": "string",
            "description": "Organization ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.SuccessResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/dto.OrganizationResponseDTO"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      }
    },
    "/organizations/{orgId}": {
      "delete": {
        "description": "Permanently delete an organization. Requires Owner role and name confirmation.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Organizations"],
        "summary": "Delete organization",
        "parameters": [
          {
            "type": "string",
            "description": "Organization ID",
            "name": "orgId",
            "in": "path",
            "required": true
          },
          {
            "description": "Deletion confirmation",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.DeleteOrganizationRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/response.SuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      },
      "patch": {
        "description": "Update organization details (name, description, slug). Requires Admin+ role.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Organizations"],
        "summary": "Update organization",
        "parameters": [
          {
            "type": "string",
            "description": "Organization ID",
            "name": "orgId",
            "in": "path",
            "required": true
          },
          {
            "description": "Updated fields",
            "name": "organization",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.UpdateOrganizationRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.SuccessResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/dto.OrganizationResponseDTO"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "409": {
            "description": "Conflict",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      }
    },
    "/organizations/{orgId}/archive": {
      "post": {
        "description": "Deactivate (archive) an organization. Requires Admin+ role.",
        "produces": ["application/json"],
        "tags": ["Organizations"],
        "summary": "Archive organization",
        "parameters": [
          {
            "type": "string",
            "description": "Organization ID",
            "name": "orgId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/response.SuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      }
    },
...
    "/organizations/{orgId}/restore": {
      "post": {
        "description": "Reactivate a previously archived organization. Requires Admin+ role.",
        "produces": ["application/json"],
        "tags": ["Organizations"],
        "summary": "Restore organization",
        "parameters": [
          {
            "type": "string",
            "description": "Organization ID",
            "name": "orgId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/response.SuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "dto.CreateOrganizationRequestDTO": {
      "type": "object",
      "required": ["name"],
      "properties": {
        "description": {
          "type": "string",
          "maxLength": 500,
          "example": "Building amazing products"
        },
        "name": {
          "type": "string",
          "maxLength": 100,
          "minLength": 3,
          "example": "Acme Corp"
        }
      }
    },
...
    "dto.DeleteOrganizationRequestDTO": {
      "type": "object",
      "required": ["confirmName"],
      "properties": {
        "confirmName": {
          "type": "string",
          "example": "Acme Corp"
        }
      }
    },
...
    "dto.OrganizationDTO": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string",
          "example": "2023-01-01T00:00:00Z"
        },
        "deactivatedAt": {
          "type": "string"
        },
        "deactivatedBy": {
          "type": "string"
        },
        "description": {
          "type": "string",
          "example": "Building amazing products"
        },
        "id": {
          "type": "string",
          "example": "org_550e8400-e29b-41d4-a716-446655440000"
        },
        "isActive": {
          "type": "boolean",
          "example": true
        },
        "maxMembers": {
          "type": "integer",
          "example": 10
        },
        "maxProjects": {
          "type": "integer",
          "example": 5
        },
        "memberCount": {
          "type": "integer",
          "example": 3
        },
        "name": {
          "type": "string",
          "example": "Acme Corp"
        },
        "ownerId": {
          "type": "string",
          "example": "user_550e8400-e29b-41d4-a716-446655440000"
        },
        "slug": {
          "type": "string",
          "example": "acme-corp"
        }
      }
    },
    "dto.OrganizationMemberDTO": {
      "type": "object",
      "properties": {
        "deactivatedAt": {
          "type": "string"
        },
        "deactivatedBy": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "example": "john@example.com"
        },
        "fullName": {
          "type": "string",
          "example": "John Doe"
        },
        "id": {
          "type": "string",
          "example": "orgmem_550e8400-e29b-41d4-a716-446655440000"
        },
        "invitedAt": {
          "type": "string",
          "example": "2023-01-01T00:00:00Z"
        },
        "invitedBy": {
          "type": "string",
          "example": "orgmem_550e8400-e29b-41d4-a716-446655440001"
        },
        "isActive": {
          "type": "boolean",
          "example": true
        },
        "joinedAt": {
          "type": "string",
          "example": "2023-01-01T00:00:00Z"
        },
        "userId": {
          "type": "string",
          "example": "user_550e8400-e29b-41d4-a716-446655440000"
        }
      }
    },
...
    "dto.OrganizationResponseDTO": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string",
          "example": "2023-01-01T00:00:00Z"
        },
        "deactivatedAt": {
          "type": "string"
        },
        "deactivatedBy": {
          "type": "string"
        },
        "description": {
          "type": "string",
          "example": "Building amazing products"
        },
        "id": {
          "type": "string",
          "example": "org_550e8400-e29b-41d4-a716-446655440000"
        },
        "isActive": {
          "type": "boolean",
          "example": true
        },
        "maxMembers": {
          "type": "integer",
          "example": 10
        },
        "maxProjects": {
          "type": "integer",
          "example": 5
        },
        "memberCount": {
          "type": "integer",
          "example": 3
        },
        "members": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/dto.OrganizationMemberDTO"
          }
        },
        "name": {
          "type": "string",
          "example": "Acme Corp"
        },
        "ownerId": {
          "type": "string",
          "example": "user_550e8400-e29b-41d4-a716-446655440000"
        },
        "slug": {
          "type": "string",
          "example": "acme-corp"
        }
      }
    },
...
    "dto.UpdateOrganizationRequestDTO": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string",
          "maxLength": 500,
          "example": "Building amazing products"
        },
        "name": {
          "type": "string",
          "maxLength": 100,
          "minLength": 3,
          "example": "Acme Corp"
        },
        "slug": {
          "type": "string",
          "maxLength": 100,
          "minLength": 3,
          "example": "acme-corp"
        }
      }
    },
...
    "response.ErrorDetail": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "example": "ERR_VALIDATION_FAILED"
        },
        "details": {},
        "message": {
          "type": "string",
          "example": "Validation failed"
        },
        "metadata": {
          "type": "object",
          "additionalProperties": true
        },
        "requestId": {
          "type": "string",
          "example": "req-123456"
        },
        "timestamp": {
          "type": "string",
          "example": "2026-02-08T18:30:00Z"
        }
      }
    },
    "response.ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/response.ErrorDetail"
        },
        "statusCode": {
          "type": "integer",
          "example": 400
        },
        "success": {
          "type": "boolean",
          "example": false
        }
      }
    },
    "response.Meta": {
      "type": "object",
      "properties": {
        "limit": {
          "type": "integer",
          "example": 10
        },
        "page": {
          "type": "integer",
          "example": 1
        },
        "requestId": {
          "type": "string",
          "example": "req-123456"
        },
        "timestamp": {
          "type": "string",
          "example": "2026-02-08T18:30:00Z"
        },
        "total": {
          "type": "integer",
          "example": 100
        },
        "totalPages": {
          "type": "integer",
          "example": 10
        }
      }
    },
    "response.SuccessResponse": {
      "type": "object",
      "properties": {
        "data": {},
        "message": {
          "type": "string",
          "example": "Operation successful"
        },
        "meta": {
          "$ref": "#/definitions/response.Meta"
        },
        "statusCode": {
          "type": "integer",
          "example": 200
        },
        "success": {
          "type": "boolean",
          "example": true
        }
      }
    }
  }
}